refactor: move the shared RBS environment out of SteepBridge
The env was never a bridge internal. `RbsDefinitionResolver` read
`SteepBridge.definition_builder` for pure RBS lookups, and `RbsTypeLookup`
documents mirroring the bridge's `Dir.pwd` guard and pairing its `reset!` with
the bridge's — two classes reaching through a type checker to get an RBS
environment.

It becomes `Signatures::SteepEnvironment`, a sibling rather than a nested
collaborator. Unlike the analyzer extractions, no delegation is left behind:
keeping `SteepBridge.definition_builder` alive would preserve exactly the
dependency this removes. The three call sites move instead
(`rbs_definition_resolver.rb`, `bin/rbs_infer`), and `RbsDefinitionResolver`
no longer depends on `SteepBridge` at all.

`steep_subtyping`/`steep_constant_resolver` stay on the bridge as private
readers over the module (0 external callers). SteepBridge is left with no
class-level API beyond the two pure `expression_key` helpers.

steep_bridge.rb 659 → 546.
