+++ lib/rbs_infer/analyzer.rb
@@ -1265,6 +1265,7 @@ require_relative "inference/return_type_resolver"
 require_relative "inference/param_type_inferrer"
 require_relative "project/source_index"
 require_relative "project/mixin_index"
+require_relative "signatures/steep_environment"
 require_relative "signatures/steep_bridge"
 require_relative "signatures/steep_bridge/lexical_scope"
 require_relative "signatures/steep_bridge/type_formatter"
